+++ views/report/index.php
@@ -1,9 +1,8 @@
 <?php
 
 use humhub\modules\reportcontent\models\ReportContent;
-use humhub\modules\ui\form\widgets\ActiveForm;
-use humhub\widgets\ModalButton;
-use humhub\widgets\ModalDialog;
+use humhub\widgets\modal\Modal;
+use humhub\widgets\modal\ModalButton;
 
 /**
  * @var $content \humhub\modules\content\models\Content
@@ -13,16 +12,12 @@ use humhub\widgets\ModalDialog;
 
 ?>
 
-<?php ModalDialog::begin(['header' => Yii::t('ReportcontentModule.base', '<strong>Report</strong> Content')]); ?>
-<?php $form = ActiveForm::begin(['id' => 'report-content-form']); ?>
-<div class="modal-body">
-    <?= $form->field($model, 'content_id')->hiddenInput()->label(false); ?>
-    <?= $form->field($model, 'comment_id')->hiddenInput()->label(false); ?>
-    <?= $form->field($model, 'reason')->radioList($model->getReasons(true)); ?>
-</div>
-<div class="modal-footer">
-    <?= ModalButton::submitModal(null, Yii::t('ReportcontentModule.base', 'Send')) ?>
-    <?= ModalButton::cancel() ?>
-</div>
-<?php ActiveForm::end() ?>
-<?php ModalDialog::end(); ?>
+<?php $form = Modal::beginFormDialog([
+    'title' => Yii::t('ReportcontentModule.base', '<strong>Report</strong> Content'),
+    'footer' => ModalButton::cancel() . ModalButton::save(Yii::t('ReportcontentModule.base', 'Send'))->submit(),
+    'form' => ['id' => 'report-content-form'],
+]); ?>
+    <?= $form->field($model, 'content_id')->hiddenInput()->label(false) ?>
+    <?= $form->field($model, 'comment_id')->hiddenInput()->label(false) ?>
+    <?= $form->field($model, 'reason')->radioList($model->getReasons(true)) ?>
+<?php Modal::endFormDialog(); ?>