+++ widgets/ReportContentLink.php
@@ -3,15 +3,15 @@
 namespace humhub\modules\reportcontent\widgets;
 
 use humhub\components\Widget;
-use humhub\libs\Html;
+use humhub\helpers\Html;
 use humhub\modules\reportcontent\helpers\Permission;
 use humhub\modules\content\components\ContentActiveRecord;
+use humhub\modules\ui\icon\widgets\Icon;
 use Yii;
 use yii\helpers\Url;
 
 class ReportContentLink extends Widget
 {
-
     /**
      * @var ContentActiveRecord
      */
@@ -25,10 +25,13 @@ class ReportContentLink extends Widget
         if (Permission::canReportContent($this->record)) {
             $reportUrl = Url::to(['/reportcontent/report', 'contentId' => $this->record->content->id]);
 
-            return Html::tag('li',
-                Html::tag('a',
-                    '<i class="fa fa-exclamation-circle"></i>' . Yii::t('ReportcontentModule.base', 'Report'),
-                    ['data-action-click' => 'ui.modal.load', 'data-action-click-url' => $reportUrl])
+            return Html::tag(
+                'li',
+                Html::a(
+                    Icon::get('exclamation-triangle') . ' ' . Yii::t('ReportcontentModule.base', 'Report'),
+                    '#',
+                    ['data-action-click' => 'ui.modal.load', 'data-action-click-url' => $reportUrl],
+                ),
             );
         }
     }