+++ docs/CHANGELOG.md
@@ -1,9 +1,15 @@
 Changelog
 =========
 
-1.2.0 (Unreleased)
+1.2.1 (Unreleased)
 ----------------------
 - Fix #81: Update module resources path
+
+1.2.0 (July 26, 2025)
+---------------------
+- Enh #84: Migration to Bootstrap 5 for HumHub 1.18
+- Enh #82: Set notification category "Administrative"
+- Enh #83: Use PHP CS Fixer
 
 1.1.0 (May 1, 2024)
 ----------------------